# [Performance] Support MQA/GQA in prefill stage by using FlashAttention