--- a/requirements.txt
+++ b/requirements.txt
@@ -6,6 +6,7 @@ numpy
 torch == 2.1.2
 transformers >= 4.36.0  # Required for Mixtral.
 xformers == 0.0.23.post1  # Required for CUDA 12.1.
+flash-attn >= 2.4.2
 fastapi
 uvicorn[standard]
 pydantic == 1.10.13  # Required for OpenAI server.