As shown in https://github.com/vllm-project/vllm/issues/1880, xformers has not supported MQA/GQA yet. So `key` and `value` need to be extended before calculating `softmax(Q @ K^T * softmax_scale) @ V`. While FlashAttention has supported MQA/GQA and has supported Turing, Ampere, Ada, or Hopper GPU. But note that FA has some limits (1. head size up to 256. datatype fp16 and bf16.) 

So for prefill, I replaced xformers with FlashAttention when FA can handle. It will fallback to xformers when head size > 256 or dtype is float. ~~Benchmark shows the speedup is 3x~~

- Using CodeLLaMA-34B config (num_query_heads=64, num_key_value_heads=8, head_size=128)
- Tested on A100-40GB
- The latency is the time for caculating softmax(Q @ K^T * softmax_scale) @ V
- The benchmark below could be reproduced by running `benchmark_multi_query_kv_attention.py`

The benchmark below is invalid as I misused `torch.repeat_interleave` when benchmarking original case.
Test id | Batchsize | Prompt length | Original xformers (us) | FA (us) | Speedup (Original / FA)
---|---|---|---|---|---
1 | 1 | 1024 | 6.875 | 1.625 | 4.2
2 | 10 | 1024 | 36.965 | 11.518 | 3.2
3 | 100 | 1024 |364.861 | 126.109 | 2.9