@Yard1 2024-01-10
Hey, I cannot replicate the benchmark results wrt current implementation. Here is what I am getting:
* `--num-query-heads 64 --num-kv-heads 8 --head-size 128  --seq-len 1024  --batch-size 1` : 2.715 us
* `--num-query-heads 64 --num-kv-heads 8 --head-size 128  --seq-len 1024  --batch-size 1`, modified to use non-blocking `repeat_interleave` (see vllm `attention.py`) : 1.876 us
* `--num-query-heads 64 --num-kv-heads 8 --head-size 128  --seq-len 1024  --batch-size 1 --use-flash-attn`: 1.722 us

* `--num-query-heads 64 --num-kv-heads 8 --head-size 128  --seq-len 1024  --batch-size 10` : 20.311 us
* `--num-query-heads 64 --num-kv-heads 8 --head-size 128  --seq-len 1024  --batch-size 10`, modified to use non-blocking `repeat_interleave` (see vllm `attention.py`) : 12.439 us
* `--num-query-heads 64 --num-kv-heads 8 --head-size 128  --seq-len 1024  --batch-size 1 --use-flash-attn`: 12.166 us

As you can see, the flash attention result is close to what you have reported, but the current implementation result is much closer to flash attention, especially after the benchmark code has been modified to match what's actually in vllm code.

Here's the modification:
```python
                key_expanded = key
                value_expanded = value
                query_expanded = query
                num_queries_per_kv = num_query_heads // num_kv_heads
                if num_queries_per_kv > 1:
                    # Handle MQA and GQA
                    query_expanded = query_expanded.view(query_expanded.shape[0], num_kv_heads,
                                    num_queries_per_kv, query_expanded.shape[-1])
                    key_expanded = key[:, :,
                            None, :].expand(key.shape[0], num_kv_heads,
                                            num_queries_per_kv, key.shape[-1])
                    value_expanded = value[:, :,
                                None, :].expand(value.shape[0], num_kv_heads,
                                                num_queries_per_kv,
                                                value.shape[-1])
                attn_bias = BlockDiagonalCausalMask.from_seqlens(seq_lens)
                output = xops.memory_efficient_attention_forward(
                    query_expanded.unsqueeze(0),
                    key_expanded.unsqueeze(0),
                    value_expanded.unsqueeze(0),
                    attn_bias=attn_bias,
                    p=0.0,
                    scale=scale,
                )
```

A100-40GB, `xformers==0.0.23.post1` `flash_attn==2.4.2`.