@zhaoyang-star 2024-01-11
@Yard1 Thanks for your info. You are right. I should use `expand` same as in attention.py, rather than `torch.repeat_interleave`. 

Updated benchmark results are below. Xformers is close to FA.

Test id | Batchsize | Prompt length | Original xformers same in attention.py (us) | FA (us) | Speedup (Original / FA)
---|---|---|---|---|---
1 | 1 | 1024 | 2.010 | 1.625 | 1.24
2 | 10 | 1024 | 13.188 | 11.518 | 1.14
3 | 100 | 1024 |140.809 | 126.109 | 1.12

xformers==0.0.22, flash-attn==2.4.2.