@zhaoyang-star 2024-01-11
> For MQA/GQA, it should mainly see a speedup during decoding, although this looks good for a start. Can we measure throughput difference?

I used starcoder which is a MQA model and the throughput is mainly close to the original version.