@sh1ng 2024-01-11
https://pytorch.org/docs/stable/generated/torch.Tensor.expand.html
>Expanding a tensor does not allocate new memory, but only creates a new view on the existing tensor where a dimension of size one is expanded to a larger size by setting the stride to 0. Any dimension of size 1 can be expanded to an arbitrary value without allocating new memory.

I guess it explains the difference.