@zhaoyang-star 2024-01-12
> > > For MQA/GQA, it should mainly see a speedup during decoding, although this looks good for a start. Can we measure throughput difference?
> > 
> > 
> > I used starcoder which is a MQA model and the throughput is mainly close to the original version.
> 
> Hi, regarding the llama structure, does it significantly improve throughput performance?

I guess the result will be similar with starcoder as starcoder is MQA and llama2-34B is GQA. Besides, llama2-7B and 13B is MHA so will not gain any speedup.