@zhaoyang-star 2024-01-13
> When I tested this PR with the llama2 70B, the throughput did not improve. I used the asyncLLMengine server along with 4*A800 80G PCIE.

Because the attention caculation could achieve 1.1~1.2 speedup, the e2e speedup is hard to benchmark.