@sighingnow 2024-02-23
See also https://github.com/vllm-project/vllm/pull/3010 which introduces the `flash_attn_with_kvcache` (available for paged kv-cache since flash-attn>=2.5.0).