@WoosukKwon 2024-08-01
Closing as it's already implemented. Thanks for submitting the PR. Learned a lot from it!